+++ app/pages/reservation/ReservationPage.js
@@ -17,7 +17,6 @@ import { fetchResource } from '../../actions/resourceActions';
 import {
   clearReservations,
   closeReservationSuccessModal,
-  openResourceTermsModal,
   setSelectedTimeSlots,
 } from '../../actions/uiActions';
 import PageWrapper from '../PageWrapper';
@@ -153,6 +152,11 @@ class UnconnectedReservationPage extends Component {
           } : {};
 
         if (isOrder) {
+          // Save billing information to respa
+          // eslint-disable-next-line no-param-reassign
+          values.reserverName = `${values.billingFirstName} ${values.billingLastName}`;
+          // eslint-disable-next-line no-param-reassign
+          values.reserverEmailAddress = values.billingEmailAddress;
           this.setState({ view: 'payment' });
         }
         allReservations.forEach(reservation => actions.postReservation({
@@ -303,7 +307,6 @@ class UnconnectedReservationPage extends Component {
                       onBack={this.handleBack}
                       onCancel={this.handleCancel}
                       onConfirm={this.handleReservation}
-                      openResourceTermsModal={actions.openResourceTermsModal}
                       reservation={reservationToEdit}
                       resource={resource}
                       selectedTime={selectedTime}
@@ -362,7 +365,6 @@ function mapDispatchToProps(dispatch) {
     clearReservations,
     closeReservationSuccessModal,
     fetchResource,
-    openResourceTermsModal,
     putReservation,
     postReservation,
     removeReservation: recurringReservationsConnector.removeReservation,