+++ app/pages/reservation/reservation-information/ReservationInformationForm.js
@@ -11,9 +11,8 @@ import constants from '../../../constants/AppConstants';
 import FormTypes from '../../../constants/FormTypes';
 import ReservationMetadataField from './ReservationMetadataField';
 import injectT from '../../../i18n/injectT';
-import ReservationTermsModal from '../../../shared/modals/reservation-terms/ReservationTermsModal';
-import PaymentTermsModal from '../../../shared/modals/payment-terms/PaymentTermsModal';
 import { hasProducts } from '../../../utils/resourceUtils';
+import WrappedText from '../../../shared/wrapped-text/WrappedText';
 import InternalReservationFields from './InternalReservationFields';
 
 const validators = {
@@ -94,17 +93,20 @@ export function validate(values, { fields, requiredFields, t }) {
 }
 
 class UnconnectedReservationInformationForm extends Component {
-  state = {
-    isPaymentTermsModalOpen: false,
-  }
-
-  closePaymentTermsModal = () => {
-    this.setState({ isPaymentTermsModalOpen: false });
-  }
-
-  openPaymentTermsModal = () => {
-    this.setState({ isPaymentTermsModalOpen: true });
-  }
+  getAsteriskExplanation = () => {
+    const { resource, t } = this.props;
+    const maybeBillable = resource.minPricePerHour && resource.maxPricePerHour;
+    if (resource.needManualConfirmation && maybeBillable) {
+      return `${t('ConfirmReservationModal.priceInfo')} ${t('ReservationForm.reservationFieldsAsteriskManualBilling')}`;
+    }
+    if (resource.needManualConfirmation && !hasProducts(resource)) {
+      return t('ReservationForm.reservationFieldsAsteriskManualBilling');
+    }
+    if (!resource.needManualConfirmation && hasProducts(resource)) {
+      return t('ReservationForm.reservationFieldsAsteriskNormal');
+    }
+    return t('ReservationForm.reservationFieldsAsteriskNormal');
+  };
 
   renderField(name, type, label, help = null) {
     if (!includes(this.props.fields, name)) {
@@ -124,17 +126,15 @@ class UnconnectedReservationInformationForm extends Component {
   }
 
   renderTermsField(name) {
-    const { openResourceTermsModal, t } = this.props;
-    const label = t('ReservationInformationForm.termsAndConditionsLabel');
-    const labelLink = `${t('ReservationInformationForm.termsAndConditionsLink')}`;
+    const { t } = this.props;
+    // eslint-disable-next-line max-len
+    const label = `${t('ReservationInformationForm.termsAndConditionsLabel')} ${t('ReservationInformationForm.termsAndConditionsLink')}`;
     return (
       <Field
         component={TermsField}
         key={name}
         label={label}
-        labelLink={labelLink}
         name={name}
-        onClick={openResourceTermsModal}
         type="terms"
       />
     );
@@ -142,14 +142,14 @@ class UnconnectedReservationInformationForm extends Component {
 
   renderPaymentTermsField = () => {
     const { t } = this.props;
+    // eslint-disable-next-line max-len
+    const label = `${t('ReservationInformationForm.paymentTermsAndConditionsLabel')} ${t('ReservationInformationForm.paymentTermsAndConditionsLink')}`;
     return (
       <Field
         component={TermsField}
         key="paymentTermsAndConditions"
-        label={t('ReservationInformationForm.paymentTermsAndConditionsLabel')}
-        labelLink={t('ReservationInformationForm.paymentTermsAndConditionsLink')}
+        label={label}
         name="paymentTermsAndConditions"
-        onClick={this.openPaymentTermsModal}
         type="terms"
       />
     );
@@ -193,17 +193,6 @@ class UnconnectedReservationInformationForm extends Component {
     );
   }
 
-  renderInfoTexts = () => {
-    const { resource, t } = this.props;
-    if (!resource.needManualConfirmation) return null;
-
-    return (
-      <div className="app-ReservationInformation__info-texts">
-        <p>{t('ConfirmReservationModal.priceInfo')}</p>
-      </div>
-    );
-  }
-
   render() {
     const {
       isEditing,
@@ -218,9 +207,6 @@ class UnconnectedReservationInformationForm extends Component {
       isStaff,
       valid,
     } = this.props;
-    const {
-      isPaymentTermsModalOpen,
-    } = this.state;
 
     this.requiredFields = staffEventSelected
       ? constants.REQUIRED_STAFF_EVENT_FIELDS
@@ -243,9 +229,8 @@ class UnconnectedReservationInformationForm extends Component {
             />
             )
           }
-          {this.renderInfoTexts()}
           <p>
-            {t('ReservationForm.reservationFieldsAsteriskExplanation')}
+            {this.getAsteriskExplanation()}
           </p>
           { includes(fields, 'reserverName') && (
             <h2 className="app-ReservationPage__title">
@@ -367,7 +352,26 @@ class UnconnectedReservationInformationForm extends Component {
             { min: '0' }
           )}
           {termsAndConditions
-            && this.renderTermsField('termsAndConditions')
+          && (
+          <React.Fragment>
+            <h2 className="app-ReservationPage__title">{t('ReservationTermsModal.resourceTermsTitle')}</h2>
+            <div className="terms-box">
+              <WrappedText text={resource.genericTerms} />
+            </div>
+            {this.renderTermsField('termsAndConditions')}
+          </React.Fragment>
+          )
+          }
+          {includes(fields, 'paymentTermsAndConditions')
+            && (
+            <React.Fragment>
+              <h2 className="app-ReservationPage__title">{t('paymentTerms.title')}</h2>
+              <div className="terms-box">
+                {t('paymentTerms.terms')}
+              </div>
+              {this.renderPaymentTermsField()}
+            </React.Fragment>
+            )
           }
           {resource.specificTerms && (
             <div>
@@ -375,9 +379,6 @@ class UnconnectedReservationInformationForm extends Component {
               <p className="specificTermsContent">{resource.specificTerms}</p>
             </div>
           )}
-          {includes(fields, 'paymentTermsAndConditions')
-            && this.renderPaymentTermsField()
-          }
           <div>
             <Button
               onClick={onCancel}
@@ -394,17 +395,12 @@ class UnconnectedReservationInformationForm extends Component {
               </Button>
               )
             }
-            {hasProducts(resource)
+            {hasProducts(resource) && !isStaff
               ? this.renderPayButton()
               : this.renderSaveButton()
             }
           </div>
         </Form>
-        <ReservationTermsModal resource={resource} />
-        <PaymentTermsModal
-          isOpen={isPaymentTermsModalOpen}
-          onDismiss={this.closePaymentTermsModal}
-        />
       </div>
     );
   }
@@ -418,7 +414,6 @@ UnconnectedReservationInformationForm.propTypes = {
   onBack: PropTypes.func.isRequired,
   onCancel: PropTypes.func.isRequired,
   onConfirm: PropTypes.func.isRequired,
-  openResourceTermsModal: PropTypes.func.isRequired,
   requiredFields: PropTypes.array.isRequired,
   resource: PropTypes.object.isRequired,
   staffEventSelected: PropTypes.bool,