+++ app/pages/user-reservations/reservation-list/ReservationListItem.js
@@ -4,14 +4,16 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import iconHome from 'hel-icons/dist/shapes/home.svg';
 
+import InfoLabel from '../../../../src/common/label/InfoLabel';
+import constants from '../../../constants/AppConstants';
 import iconCalendar from '../../../assets/icons/calendar.svg';
 import ReservationAccessCode from '../../../shared/reservation-access-code/ReservationAccessCode';
 import ReservationControls from '../../../shared/reservation-controls/ReservationControlsContainer';
-import ReservationStateLabel from '../../../shared/reservation-state-label/ReservationStateLabel';
 import TimeRange from '../../../shared/time-range/TimeRange';
 import injectT from '../../../i18n/injectT';
 import { getMainImage } from '../../../utils/imageUtils';
-import { getResourcePageUrl } from '../../../utils/resourceUtils';
+import { getResourcePageUrl, hasProducts } from '../../../utils/resourceUtils';
+import { getReservationPrice, getTaxPercentage } from '../../../../src/domain/resource/utils';
 
 class ReservationListItem extends Component {
   renderImage(image) {
@@ -28,6 +30,16 @@ class ReservationListItem extends Component {
 
     const nameSeparator = isEmpty(resource) || isEmpty(unit) ? '' : ', ';
 
+    const price = getReservationPrice(reservation.begin, reservation.end, resource);
+    const vat = getTaxPercentage(resource);
+    const tVariables = {
+      price,
+      vat
+    };
+
+    const paymentLabel = constants.RESERVATION_PAYMENT_LABELS[reservation.state];
+    const statusLabel = constants.RESERVATION_STATE_LABELS[reservation.state];
+
     return (
       <li className="reservation">
         <div className="col-md-3 col-lg-2 image-container">
@@ -36,7 +48,14 @@ class ReservationListItem extends Component {
           </Link>
         </div>
         <div className="col-xs-8 col-md-6 col-lg-7 reservation-details">
-          <ReservationStateLabel reservation={reservation} />
+          <div className="reservation-state-label-container">
+            {hasProducts(resource)
+              && !reservation.staffEvent
+              && price > 0 && (
+                <InfoLabel labelStyle={paymentLabel.labelBsStyle} labelText={t(paymentLabel.labelTextId)} />
+            )}
+            <InfoLabel labelStyle={statusLabel.labelBsStyle} labelText={t(statusLabel.labelTextId)} />
+          </div>
           <Link to={getResourcePageUrl(resource)}>
             <h4>{resource.name}</h4>
           </Link>
@@ -55,6 +74,14 @@ class ReservationListItem extends Component {
             resource={resource}
             text={t('ReservationListItem.accessCodeText')}
           />
+          {hasProducts(resource)
+            && !reservation.staffEvent
+            && price > 0 && (
+            <div>
+              <span className="price">{`${t('common.totalPriceLabel')}: `}</span>
+              <span>{t('common.priceWithVAT', tVariables)}</span>
+            </div>
+          )}
         </div>
         <div className="col-xs-4 col-md-3 col-lg-3 action-container">
           <ReservationControls