+++ app/pages/user-reservations/reservation-list/__tests__/ReservationListContainer.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import Immutable from 'seamless-immutable';
+import simple from 'simple-mock';
 
 import Reservation from '../../../../utils/fixtures/Reservation';
 import Resource from '../../../../utils/fixtures/Resource';
@@ -11,13 +12,20 @@ import {
 import ReservationListItem from '../ReservationListItem';
 
 describe('pages/user-reservations/reservation-list/ReservationListContainer', () => {
+  const fetchReservations = simple.stub();
   function getWrapper(props) {
     const defaults = {
+      fetchReservations,
       isAdmin: false,
       loading: false,
+      location: {
+        search: ''
+      },
       reservations: [],
       resources: {},
       staffUnits: [],
+      paginatedReservations: {},
+      pageSize: 0,
       units: {},
     };
     return shallowWithIntl(<ReservationListContainer {...defaults} {...props} />);
@@ -28,10 +36,25 @@ describe('pages/user-reservations/reservation-list/ReservationListContainer', ()
     const resource = Resource.build({ unit: unit.id });
     const props = {
       isAdmin: false,
+      location: {
+        search: '',
+      },
       reservations: Immutable([
         Reservation.build({ resource: resource.id }),
         Reservation.build({ resource: 'unfetched-resource' }),
       ]),
+      paginatedReservations: {
+        count: 0,
+        comingReservations: Immutable([
+          Reservation.build({ resource: resource.id }),
+          Reservation.build({ resource: 'unfetched-resource' }),
+        ]),
+        pastReservations: Immutable([
+          Reservation.build({ resource: resource.id }),
+          Reservation.build({ resource: 'unfetched-resource' }),
+        ]),
+      },
+      pageSize: 10,
       resources: Immutable({
         [resource.id]: resource,
       }),
@@ -52,16 +75,16 @@ describe('pages/user-reservations/reservation-list/ReservationListContainer', ()
     describe('rendering individual reservations', () => {
       test('renders a ReservationListItem for every reservation in props', () => {
         const reservationListItems = getWithReservationsWrapper().find(ReservationListItem);
-        expect(reservationListItems).toHaveLength(props.reservations.length);
+        const { comingReservations, pastReservations } = props.paginatedReservations;
+        expect(reservationListItems).toHaveLength(comingReservations.length + pastReservations.length);
       });
 
       test('passes isAdmin, isStaff and reservation', () => {
         const reservationListItems = getWithReservationsWrapper().find(ReservationListItem);
-        reservationListItems.forEach((reservationListItem, index) => {
+        reservationListItems.forEach((reservationListItem) => {
           const actualProps = reservationListItem.props();
           expect(actualProps.isAdmin).toBe(props.isAdmin);
           expect(actualProps.isStaff).toBe(false);
-          expect(actualProps.reservation).toEqual(props.reservations[index]);
           expect(reservationListItems.at(0).prop('resource')).toEqual(resource);
           expect(reservationListItems.at(1).prop('resource')).toEqual({});
           expect(reservationListItems.at(0).prop('unit')).toEqual(unit);