+++ app/shared/availability-view/timeline-groups/timeline-group/TimelineGroup.js
@@ -83,6 +83,7 @@ export default class TimelineGroup extends React.Component {
       onSelectionCancel,
       selection,
     } = this.props;
+    const selectionDuration = selection ? moment(selection.end).diff(selection.begin, 'hours') : null;
     return (
       <div
         className={classNames('timeline-group', this.props.className)}
@@ -101,6 +102,9 @@ export default class TimelineGroup extends React.Component {
                 className={classNames('hour', {
                   'hour-start-selected': selection && range.midPoint.isSame(selection.end),
                   'hour-end-selected': selection && range.endTime.isSame(selection.end),
+                  'hour-whole-selected': selection
+                    && selectionDuration > 0
+                    && moment(range.midPoint).isBetween(selection.begin, selection.end),
                 })}
                 key={range.startTime.format('HH')}
                 style={{ width: utils.getTimeSlotWidth(range) }}