+++ app/shared/availability-view/timeline-groups/timeline-group/availability-timeline/AvailabilityTimeline.js
@@ -1,10 +1,13 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import { connect } from 'react-redux';
+import { createSelector } from 'reselect';
 
 import Reservation from './Reservation';
 import ReservationSlot from './ReservationSlot';
+import { resourcesSelector } from '../../../../../state/selectors/dataSelectors';
 
-export default class AvailabilityTimeline extends React.Component {
+class UnconnectedAvailabilityTimeline extends React.Component {
   static propTypes = {
     id: PropTypes.string.isRequired,
     items: PropTypes.arrayOf(
@@ -19,7 +22,9 @@ export default class AvailabilityTimeline extends React.Component {
     onReservationSlotMouseEnter: PropTypes.func,
     onReservationSlotMouseLeave: PropTypes.func,
     onSelectionCancel: PropTypes.func,
+    products: PropTypes.array,
     selection: PropTypes.object,
+    slotSize: PropTypes.string
   };
 
   shouldComponentUpdate(nextProps) {
@@ -35,15 +40,18 @@ export default class AvailabilityTimeline extends React.Component {
       onSelectionCancel,
       onReservationSlotMouseEnter,
       onReservationSlotMouseLeave,
+      products,
       selection,
+      slotSize
     } = this.props;
     return (
       <div className="availability-timeline">
-        {this.props.items.map((item) => {
+        {this.props.items.map((item, index) => {
           if (item.type === 'reservation-slot') {
             return (
               <ReservationSlot
                 {...item.data}
+                itemIndex={index}
                 key={item.key}
                 onClick={onReservationSlotClick}
                 onMouseEnter={onReservationSlotMouseEnter}
@@ -51,6 +59,7 @@ export default class AvailabilityTimeline extends React.Component {
                 onSelectionCancel={onSelectionCancel}
                 resourceId={this.props.id}
                 selection={selection}
+                slotSize={slotSize}
               />
             );
           }
@@ -59,6 +68,7 @@ export default class AvailabilityTimeline extends React.Component {
               {...item.data}
               key={item.key}
               onClick={onReservationClick}
+              products={products}
             />
           );
         })}
@@ -66,3 +76,24 @@ export default class AvailabilityTimeline extends React.Component {
     );
   }
 }
+
+export function selector() {
+  function idSelector(state, props) {
+    return props.id;
+  }
+  const resourceSelector = createSelector(
+    resourcesSelector,
+    idSelector,
+    (resources, id) => resources[id]
+  );
+  return createSelector(
+    resourceSelector,
+    resource => ({
+      products: resource.products ? resource.products : [],
+      slotSize: resource.slotSize
+    })
+  );
+}
+export { UnconnectedAvailabilityTimeline };
+const AvailabilityTimeline = connect(selector)(UnconnectedAvailabilityTimeline);
+export default AvailabilityTimeline;